All tests pass and TF_Autoencoder.py replicates the Torch version.
